% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask}
\alias{ask}
\title{Ask a question to a chatbot}
\usage{
ask(
  question,
  context = NULL,
  max_tokens = getOption("SciViews.chatbot.max_tokens", 1000),
  lang = getOption("data.io_lang", "en"),
  url = getOption("SciViews.chatbot.url", "http://localhost:11434/api/chat"),
  model = getOption("SciViews.chatbot.model", "mistral:7b-instruct-v0.2-q6_K"),
  verbose = FALSE
)
}
\arguments{
\item{question}{A character string with the question to ask.}

\item{context}{An R object used as context (usually a data frame).}

\item{max_tokens}{The maximum number of tokens to return in the answer. By
default, it is 1000.}

\item{lang}{The language to use for the answer. Default is "en". You can also
use "fr".}

\item{url}{The URL of the chatbot server. Default is
http://localhost:11434/api/chat}

\item{model}{The LLM (large language model) to use. Default is a variant of
Mistral 7B.}

\item{verbose}{Should more information be printed? \code{FALSE} by default.}
}
\value{
The answer is returned invisibly. The function is used for its side-effect of
displaying the chatbot help page with the question, answer and examples
}
\description{
The question is sent to the chatbot server, and the answer is returned. The
default chatbot server is ollama running locally, and the default model is
mistral:7b-instruct-v0.2-q6_K
}
\examples{
\dontrun{
ask("What is a chatbot?")
ask("Write R code to filter a data frame.")
}
}
