% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai_ask.R
\name{chatbot1}
\alias{chatbot1}
\title{SciViews chatbot}
\description{
\emph{A chatbot does not always provide reliable results. Take this with a grain of salt!}

Question: \Sexpr[eval=TRUE,stage=render,results=rd]{ if (exists('SciViews.chatbot.page1')) SciViews.chatbot.page1$question else "No question" }
}
\section{Answer}{

\Sexpr[eval=TRUE,stage=render,results=rd]{ if (exists('SciViews.chatbot.page1')) SciViews.chatbot.page1$answer else "No answer" }
}

\examples{
\Sexpr[eval=TRUE,stage=render,results=text]{ if (exists('SciViews.chatbot.page1')) SciViews.chatbot.page1$examples else "# No example" }
}
