% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{insertSemicolonAddin}
\alias{insertSemicolonAddin}
\title{Insert semicolon.}
\usage{
insertSemicolonAddin()
}
\description{
Call this function as an addin to insert \code{:} at the cursor position.
This addin can be assigned to the \code{:} key in order to get rid of a bug in
RStudio server under chromium, which sometimes inserts \code{/} in place of \code{:}.
}
