% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViews_packages.R
\name{SciViews_packages}
\alias{SciViews_packages}
\alias{SciViews_conflicts}
\alias{print.SciViews_conflicts}
\title{Give the list of SciViews::R packages and check for conflicts}
\usage{
SciViews_packages(..., all = FALSE)

SciViews_conflicts(all = TRUE)

\method{print}{SciViews_conflicts}(x, ..., startup = FALSE)
}
\arguments{
\item{...}{Further topics to consider in SciViews::R (not used yet).}

\item{all}{Should all packages be listed (\code{TRUE}) or only those that are
attached to the search path (\code{FALSE}).}

\item{x}{A SciViews_conflicts object}

\item{startup}{Should the message be printed at startup?}
}
\value{
A list of packages for \code{\link[=SciViews_packages]{SciViews_packages()}}, or a
\strong{SciViews_conflicted} object with a \code{\link[=print]{print()}} method for
\code{\link[=SciViews_conflicts]{SciViews_conflicts()}}.
}
\description{
List required packages or conflicting functions. These functions
are inspired by \code{\link[tidyverse:tidyverse_packages]{tidyverse::tidyverse_packages()}} and
\code{\link[tidyverse:tidyverse_conflicts]{tidyverse::tidyverse_conflicts()}}, but adapted to the SciViews::R context.
}
\examples{
SciViews_packages()
# More complete list of packages
SciViews_packages(all = TRUE)
# Conflicts
SciViews_conflicts()
}
