% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViews-R.R
\name{SciViews_R}
\alias{SciViews_R}
\alias{R}
\alias{print.SciViews_R}
\title{Configure \R for the SciViews::R dialect}
\usage{
R(..., dtx = NULL, silent = FALSE)

\method{print}{SciViews_R}(x, ...)
}
\arguments{
\item{...}{Further parameters to configure \R (not used yet).}

\item{dtx}{Which dtx object to use be default? \code{"dtt"} or \code{"data.table"} for
data.table, \code{"dtf"} or \code{"data.frame"} for data.frame, \code{"dtbl"}, \code{"tibble"}
or \code{"tbl_df"} for tibble's tbl_df, the name of a function to use to convert
a data.frame object, or \code{NULL} (by default) to keep current settings.}

\item{silent}{If \code{TRUE}, no report is printed about loaded packages and
conflicts.}

\item{x}{An object to print.}
}
\description{
Load required packages like data.table, collapse, ggplot2, dplyr, svMisc, ...
to get a fully functional \code{SciViews::R} dialect environment.
}
\note{
Use \code{SciViews::R} instruction in the beginning of an R script, or in
the setup or first chunk of an R Markdown/Notebook to ensure the SciViews::R
dialect is correctly installed. The report indicating attached packages and
conflicts is largely inspired by the corresponding \code{tidyverse} code,
written by Hadley Wickham.
}
\examples{
\dontrun{
SciViews::R
}
}
\seealso{
\code{\link[=library]{library()}}, \code{\link[utils:install.packages]{utils::install.packages()}}
}
\concept{packages loading}
\keyword{utilities}
