\name{snippets}
\Rdversion{2}
\alias{snippets}
\docType{package}

\title{SciViews snippet help}

\description{
  \Sexpr[results=rd, stage=render]{c("We are now \\\\emph{", Sys.time(),
    "}", sep = "")}
}

\arguments{
  \Sexpr[results=rd, stage=render]{c("The content of the arguments section...", sep = "")}
}

\keyword{ package }

\concept{ snippets and electrnoic reference card }
