% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViews-package.R
\docType{package}
\name{tabularise-package}
\alias{tabularise-package}
\title{'SciViews::R' Dialect for Data Processing and Visualization}
\description{
The \code{SciViews::R} dialect is base R + tidyverse + a series of additional
SciViews packages like data.io, svBase, svFlow, tabularise or chart.
}
\section{Important functions}{

\itemize{
\item \code{\link[=R]{R()}} for loading the `SciViews::R`` packages,
\item \code{\link[=pcomp]{pcomp()}} for a PCA analysis (unifying various methods),
\item \code{\link[=correlation]{correlation()}} to calculate and plot a correlation matrix,
\item \code{\link[=panel_reg]{panel_reg()}} and others to plot panels in \code{pairs} or \code{coplot} graphs,
\item \code{\link[=panel_boxplot]{panel_boxplot()}} and others for univariate panels in \code{pairs} plots.
\item \code{\link[=rwb_colors]{rwb_colors()}} and others to generate color palettes.
\item \code{\link[=enum]{enum()}} to enumerate items in a vector,
\item \code{\link[=timing]{timing()}} to determine the time required to run an R expression,
\item \code{\link[=nr]{nr()}} and co as convenient shorthand to columns and rows,
\item \code{\link[=ln]{ln()}} and others for natural logarithm.
}
}

