% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\docType{data}
\name{ln}
\alias{ln}
\alias{lg}
\alias{lb}
\alias{ln1p}
\alias{lg1p}
\alias{E}
\title{Logarithms and exponentials}
\format{An object of class \code{numeric} of length 1.}
\usage{
ln(x)

lg(x)

lb(x)

ln1p(x)

lg1p(x)

E
}
\arguments{
\item{x}{A numeric or complex vector.}
}
\description{
\code{ln()} computes natural logarithm, \code{lg()} computes base 10 logarithm, and
\code{lb()} computes binary (base 2) logarithm.

\code{ln1p()} and \code{lg1p()} computes \code{ln(x + 1)} and \code{lg(x + 1)} accurately also
for \code{|x| << 1}.

\code{E} is the Euler constant and is equal to \code{exp(1)}.
}
\details{
Those functions are synonyms of \code{log()}, \code{log10()}, \code{log2()},
\code{log1p()} for those who prefer the shorter notation. BeginneRs sometimes
make confusion between \code{log()} and \code{log10()}. Using \code{ln()} for natural
logarithms instead of \code{log()} eliminates this confusion. \code{E} is provided for
convenience as \code{exp(1)}, although the use of \code{exp()} is usually familar
enough to everyone.
}
\examples{
ln(exp(3))              # Same as log(exp(3))
lg(10^3)                # Same as log10(10^3)
lb(1:3)                 # Wrapper for log2()

ln1p(c(0, 1, 10, 100))  # Wrapper for log1p()
lg1p(c(0, 1, 10, 100))  # log10(x + 1), but optimized for x << 1

E^4                     # Similar to exp(4), but different calculation!
}
\seealso{
\code{\link[=log]{log()}}
}
\concept{
logarithms and exponentials
}
\keyword{datasets}
\keyword{math}
