% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViews-package.R
\docType{package}
\name{SciViews-package}
\alias{SciViews-package}
\title{SciViews - Main package}
\description{
The SciViews package provides various functions to install the SciViews::R
dialect. It also provides additional utilites besides base, recommended and
tidyverse.
}
\section{Important functions}{


TODO...
}

