\name{SciViews-package}
\alias{SciViews-package}
\alias{SciViews}
\docType{package}

\title{ SciViews GUI API - Main package }

\description{
  Functions to install SciViews additions to R, and miscellaneous
}

\details{
  \tabular{ll}{
    Package: \tab SciViews\cr
    Type: \tab Package\cr
    Version: \tab 0.9-2\cr
    Date: \tab 2010-09-26\cr
    License: \tab GPL (>= 2)\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Philippe Grosjean

  Maintainer: Philippe Grosjean <phgrosjean@sciviews.org>
}

\references{
  SciViews: http://www.sciviews.org/
}

\keyword{ package }

\concept{ R enhancement }
