% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciViews-R.R
\name{SciViews_R}
\alias{SciViews_R}
\alias{R}
\alias{print.SciViews_R}
\title{Configure \R for the SciViews::R dialect}
\usage{
R(..., lang = NULL, dtx = NULL, threads.percent = 75, silent = TRUE)

\method{print}{SciViews_R}(x, ...)
}
\arguments{
\item{...}{Further topics to include to configure \R (load more packages).
Currently, \code{"infer"}, \code{"model"}, \code{"explore"}, \code{"ml"}, \code{"ts"} or \code{"spatial"}}

\item{lang}{What is the default natural language to use, e.g., \code{"en"} or
\code{"fr"}, with uppercase versions \code{"EN"} or \code{"FR"} convert even more strings,
for instance, \code{\link[data.io:read]{data.io::read()}} does not convert factor levels in the
corresponding language for supported data sets unless the uppercase version
is specified. If \code{NULL} (by default), current configuration is not changed.}

\item{dtx}{Which dtx object is to be used be default? \code{"dtt"} or
\code{"data.table"} for data.table, \code{"dtf"} or \code{"data.frame"} for data.frame,
\code{"dtbl"}, \code{"tibble"} or \code{"tbl_df"} for tibble's tbl_df, the name of a
function to use to convert a data.frame object, or \code{NULL} (by default) to
keep current settings.}

\item{threads.percent}{The percentage of threads to use for \{data.table\} and
\{collapse\} parallel code (number of threads depend on how many are
available, and the value is rounded towards zero).}

\item{silent}{If \code{TRUE} (by default), no report is printed about loaded
packages and conflicts.}

\item{x}{An object to print.}
}
\description{
Load required packages like data.table, collapse, ggplot2, dplyr, svMisc, ...
to get a fully functional \code{SciViews::R} dialect environment.
}
\note{
Use \code{SciViews::R} instruction in the beginning of an R script, or in
the setup or first chunk of an R Markdown/Notebook to ensure the SciViews::R
dialect is correctly installed. The report indicating attached packages and
conflicts is largely inspired by the corresponding \code{tidyverse} code,
written by Hadley Wickham.
}
\examples{
\dontrun{
SciViews::R
}
}
\seealso{
\code{\link[=library]{library()}}, \code{\link[utils:install.packages]{utils::install.packages()}}
}
\concept{packages loading}
\keyword{utilities}
